+++ tests/services/test_OximetryService.cpp
@@ -490,6 +490,7 @@ public:
 class MockDatabase : public IDatabase {
 public:
     HMS_CPAP_STUB_EQUIPMENT_METHODS
+    HMS_CPAP_STUB_SYNC_FOLDER_METHODS
     DbType dbType() const override { return DbType::SQLITE; }
     MOCK_METHOD(bool, connect, (), (override));
     MOCK_METHOD(void, disconnect, (), (override));